Better determine when to use provisioned vs. requested resources
Resources like CPUs and GPUs are not usage-tracked, like memory
and disk are, and so to compute stats for non-tracked resources,
we must deploy a strategy of looking at requested and/or provided
resources. When a job is matched with a machine with partitionable
slots, generally the provisioned resources will match the
requested resources (which could be expressions) and it's safe to
use the provisioned resource values to compute usage statistics.
However, not all machines use partitionable slots. A static slot
will provide the entire slot's resources, even if a job did not
request them. Therefore, we should not assume that a job will use
the resources it was provided, but we should assume that, when they
are defined, a job does use the resources it requested.

These changes look at RequestCpus and RequestGpus. If they are
defined but are expressions (i.e. are not numeric), then the script
falls back to looking at CpusProvided and GpusProvided, if they are
defined and numeric. Otherwise, assume 1 CPUs and 0 GPUs were used.

Also, don't just ignore RequestMemory and RequestDisk, as we want
to know exactly what users' jobs are requesting.