Report and use requested and provisioned resource attrs (#17)
- Reports the original RequestMemory and RequestDisk. (Doesn't seem like CHTC even reports RequestMemory_RAW or RequestDisk_RAW).

- Standardizes to camel-casing for GPU-related attributes (e.g. RequestGpus and GpusProvisioned).

- Uses/fallbacks to CpusProvisioned and GpusProvisioned when it makes sense to do so when computing usage statistics for CPUs and GPUs.

Resources like CPUs and GPUs are not usage-tracked, like memory
and disk are, and so to compute stats for non-tracked resources,
we must deploy a strategy of looking at requested and/or provided
resources. When a job is matched with a machine with partitionable
slots, generally the provisioned resources will match the
requested resources (which could be expressions) and it's safe to
use the provisioned resource values to compute usage statistics.
However, not all machines use partitionable slots. A static slot
will provide the entire slot's resources, even if a job did not
request them. Therefore, we should not assume that a job will use
the resources it was provided, but we should assume that, when they
are defined, a job does use the resources it requested.

These changes look at RequestCpus and RequestGpus. If they are
defined but are expressions (i.e. are not numeric), then the script
falls back to looking at CpusProvided and GpusProvided, if they are
defined and numeric. Otherwise, assume 1 CPUs and 0 GPUs were used.