Initial attempt at polling startd histories
1. Added configuration (process_startd_history, es_feed_startd_history)
2. Added get_startds function to get startd address ads
3. Added process_startd function and put call in process_histories

The ugliest part of this is that the same checkpoint file is used for
both schedds and startds. This will cause problems in the case where a
schedd and a startd reside on the same machine (and report to the same
collector). However, the checkpoint file is currently hardcoded in
process_histories, and it will take some major refactoring to make it
(yet another) commandline parameter.